// src/showtime/dto/response.artistsong.dto.ts
export class ResponseArtistSongDto {

  albumImage: string;

  name: string;

  artist: string;

  title: string;

  totalAmount: number;

  leftAmount: number;

  releaseStartAt: string;

  releaseEndAt: string;

  isAvailable: boolean;

  nftMusicId: number;

  musicFileUrl: string;

  tokenId: String;

  playTime: number;

  lyrics: string;
}